pyproject.toml
[tool.poetry]
name = "cheolsu-ai"
version = "0.1.0"
description = ""
authors = ["jjooki <piy8117982@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.11,<3.13"
openai = "^1.47.1"
langchain = "^0.3.0"
pandas = "^2.2.3"
fastapi = "^0.115.0"
ddtrace = "^2.14.0"
langsmith = "^0.1.129"
pymongo = "^4.9.1"
pymysql = "^1.1.1"
bcrypt = "^4.2.0"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
uvicorn = "^0.31.0"
nest-asyncio = "^1.6.0"
python-dotenv = "^1.0.1"
aiomysql = "^0.2.0"
pillow = "^10.4.0"
boto3 = "^1.35.40"
pytz = "^2024.2"
redis-om = "^0.3.2"
asyncer = "^0.0.8"


[tool.poetry.group.dev.dependencies]
ipykernel = "^6.29.5"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
